feat(categories): externalize helpers functions
