+++ src/ducks/categories/Categories.jsx
@@ -12,6 +12,7 @@ import { ConnectedSelectDates as SelectDates } from 'components/SelectDates'
 import styles from './styles'
 import CategoriesChart from './CategoriesChart'
 import { flowRight as compose } from 'lodash'
+import { getTransactionsTotal, getGlobalCurrency } from './helpers'
 
 const stAmount = styles['bnk-table-amount']
 const stCategory = styles['bnk-table-category-category']
@@ -27,14 +28,6 @@ const stUncollapsed = styles['bnk-table-row--uncollapsed']
 const stCatTotalMobile = styles['bnk-category-total-mobile']
 
 const vHidden = { visibility: 'hidden' }
-
-const subCategorySort = (a, b) => {
-  if (b.percentage !== a.percentage) {
-    return b.percentage - a.percentage
-  } else {
-    return a.amount - b.amount
-  }
-}
 
 class Categories extends Component {
   toggle = categoryName => {
@@ -62,36 +55,9 @@ class Categories extends Component {
         categories = []
       }
     }
-    let transactionsTotal = 0
-    const globalCurrency = categories.length > 0 ? categories[0].currency : '€'
 
-    if (categories.length !== 0) {
-      // compute some global data
-      const absoluteTransactionsTotal = categories.reduce(
-        (total, category) => total + Math.abs(category.amount),
-        0
-      )
-      for (let category of categories) {
-        category.percentage = Math.round(
-          Math.abs(category.amount) / absoluteTransactionsTotal * 100
-        )
-        const absoluteSubcategoriesTotal = category.subcategories.reduce(
-          (total, category) => total + Math.abs(category.amount),
-          0
-        )
-        for (let subcategory of category.subcategories) {
-          if (absoluteSubcategoriesTotal === 0) {
-            subcategory.percentage = 100
-          } else {
-            subcategory.percentage = Math.round(
-              Math.abs(subcategory.amount) / absoluteSubcategoriesTotal * 100
-            )
-          }
-        }
-        category.subcategories = category.subcategories.sort(subCategorySort)
-        transactionsTotal += category.amount
-      }
-    }
+    const globalCurrency = getGlobalCurrency(categories)
+    const transactionsTotal = getTransactionsTotal(categories)
 
     // sort the categories for display
     categories = categories.sort((a, b) => {
